+++ Chess/chessboardutilities.h
@@ -3,13 +3,15 @@
 
 #include <QObject>
 #include <QColor>
+#include "tileback.h"
+#include "figureback.h"
 
 class ChessBoardUtilities : public QObject
 {
     Q_OBJECT
     Q_PROPERTY(QColor color READ color WRITE setColor NOTIFY colorChanged)
-    Q_PROPERTY(short xCordCalculated READ xCordCalculated NOTIFY xCordCalculatedChanged)
-    Q_PROPERTY(short yCordCalculated READ yCordCalculated NOTIFY yCordCalculatedChanged)
+    Q_PROPERTY(int xCordCalculated READ xCordCalculated NOTIFY xCordCalculatedChanged)
+    Q_PROPERTY(int yCordCalculated READ yCordCalculated NOTIFY yCordCalculatedChanged)
 
 private:
     // setting tile color variables
@@ -17,8 +19,8 @@ private:
     int counterColor;
 
     //setting tile cords variables
-    short xCounter;
-    short yCounter;
+    int xCounter;
+    int yCounter;
 
 public:
     explicit ChessBoardUtilities(QObject *parent = nullptr);
@@ -26,8 +28,8 @@ public:
     QColor color();
     void setColor(const QColor &colorGiven);
 
-    short xCordCalculated();
-    short yCordCalculated();
+    int xCordCalculated();
+    int yCordCalculated();
 
 signals:
     void colorChanged();
